[CS] add TODOs and update README.md with proposed endpoints
